requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
sentence-transformers
faiss-cpu
numpy
scikit-learn
